Install overlay_map.dtb when using in-tree DT overlays
Unlike the rest of the overlays directory, this file has .dtb suffix,
which results for it not being installed when copying *.dtbo wildcard.
It is necessary for correct mapping of overlays without RPi version
suffix to the platform-specific ones.

Also use a foreach loop for installing the overlays to set the mode to
0644, just like when rpi-firmware package installs them.

Signed-off-by: Jan Čermák <sairon@sairon.cz>